Reduce repo_checker to check a project (and rename)
